refactor: vite.config.ts에 작성된 사전 렌더링 경로 생성 함수 분리
