+++ tsconfig.node.json
@@ -9,5 +9,5 @@
     "strict": true,
     "noEmit": true
   },
-  "include": ["vite.config.ts"]
+  "include": ["vite.config.ts", "src/utils/generatePerformanceRoute.ts"]
 }
